Split NodeInfo in a component folder; hide params in nodeinfo
